+++ src/pages/RentalManagement/SerializedAsset/index.tsx
@@ -15,7 +15,7 @@ import ConfirmationDialog from "../../../components/Helpers/ConfirmationDialog";
 import CustomReactTable from "../../../components/CustomReactTable/CustomReactTable";
 import ManagePurchaseOrder from "../../PurchaseOrder/ManagePurchaseOrder";
 import ManageSublease from "../../Sublease/ManageSublease";
-import { uniqBy, uniq } from 'lodash';
+import { uniqBy, uniq, startCase } from 'lodash';
 import { CustomOfflineContext } from "../../../StateProvider/OfflineContext/OfflineContext";
 import { objectStore, findOne } from '../../../constants/indexdbhelper';
 import HtmlTooltip from "../../../components/CustomTooltipTitle";
@@ -35,7 +35,8 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
   const [isAdding, setAdding] = useState(false)
   const [showConfirmBox, setShowConfirmBox] = useState(false)
   const [addSerializedAssetDialog, setAddSerializedAssetDialog] = useState({ open: false })
-  const [selectedProducts, setSelectedProducts] = useState([])
+  const [selectedRecords, setSelectedRecords] = useState([])
+
   const [assetAssignedProduct, setAssetAssignedProduct] = useState([])
 
   const [deleteData, setDeleteData] = useState([])
@@ -75,10 +76,15 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
         Cell: ({ row }) => (
           <div className="d-flex gap-2 align-items-center">
             <p className="text-truncate" title={row.original.detail}  >
-              {(row.original?.type === "asset" && !isOffline) ?
-                <a className="link text-truncate" href={`${serializedAsset.route}/detail/${row.original.inventory}`} target="_blank">{row.original.detail}</a> :
-                row.original.detail}
+              {(!isOffline) ?
+                row.original?.type === "product" ?
+                  <a className="link text-truncate" href={`${routes.productDetail.path}/${row.original.materialId}`} target="_blank">{row.original.detail}</a>
+                  : row.original?.type === "package" ?
+                    <a className="link text-truncate" href={`${routes.packagesDetail.path}/${row.original.materialId}`} target="_blank">{row.original.detail}</a>
+                    : <a className="link text-truncate" href={`${routes.serializedAssetDetail.path}/${row.original.inventory}`} target="_blank">{row.original.detail}</a>
+                : row.original.detail}
             </p>
+            <Chip className="ml-1" label={startCase(row.original?.type)} size="small" color="primary" />
             {row.original.isPurchaseOrder &&
               <HtmlTooltip title={`${routes.purchaseOrder.title}`}>
                 <IconButton size="small" onClick={() => {
@@ -103,7 +109,6 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
             }
             {row.original?.type === "asset" &&
               <span className="d-flex align-items-center gap-2">
-                <Chip label="Asset" size="small" color="primary" />
                 {(row.original.status === INVENTORY_STATUS.reserved && row.original?.manualStatus !== INVENTORY_STATUS.reserved && !isOffline) &&
                   <HtmlTooltip title={`Remove`}>
                     <IconButton size="small" onClick={() => {
@@ -255,88 +260,9 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
         parent.realAssetAssignedQty = parent.assetAssignedQty;
         parent.hideSelection = parent.type === "product" && !parent.productDetail?.serializedProduct ? true : false;
         parent.isValid = parent.serializedProduct ? parent.assetAssignedQty === parent.assetQty ? true : false : true;
-
         parent.isSublease = subleaseProduct?.some(e => e.materialId === parent.materialId)
         parent.isPurchaseOrder = purchaseOrderProduct?.some(e => e.productId === parent.materialId)
-
         parent.subRows = generateNestedData(data.material, data.inventory, parent, transferAssets, subleaseProduct, purchaseOrderProduct);
-
-        // const subRows = []
-        // const inventory = data.inventory?.filter((e) => e._id === parent._id);
-        // inventory?.forEach((_inventory, k) => {
-        //   const transferFilter = transferAssets.filter(e => e.assetId === _inventory.inventory);
-        //   var isTransferAsset = false;
-        //   var transferData = {};
-        //   if (transferFilter.length) {
-        //     isTransferAsset = true
-        //     transferData = transferFilter[0]
-        //   }
-        //   const isPurchaseOrderAsset = purchaseOrderProduct.some(e => e._id === _inventory?.inventoryDetail?.purchaseOrder);
-        //   subRows.push({
-        //     ..._inventory,
-        //     srno: `${(i + 1)}.${(k + 1)}`,
-        //     detail: _inventory.inventoryDetail?.assetNumber,
-        //     type: "asset",
-        //     status: _inventory.inventoryDetail?.status,
-        //     manualStatus: _inventory.inventoryDetail?.manualStatus,
-        //     _id: _inventory.inventory,
-        //     isValid: _inventory.inventoryDetail?.manualStatus === INVENTORY_STATUS.reserved ? false : true,
-        //     isPurchaseOrderAsset: isPurchaseOrderAsset,
-        //     isTransferAsset: isTransferAsset,
-        //     transferData: transferData,
-        //     isSubleaseAsset: _inventory.inventoryDetail?.subleaseAsset
-        //   })
-        // })
-        // parent.subRows = subRows;
-        // parent.isSublease = subleaseProduct?.some(e => e.materialId === parent.materialId)
-        // if (parent.type === "product") {
-        //   parent.isValid = parent?.qty === subRows?.length ? true : false;
-        //   parent.isPurchaseOrder = purchaseOrderProduct?.some(e => e.productId === parent.materialId)
-        // }
-        // if (parent.type === "package") {
-        //   const child: any = [...data.material?.filter((e) => e.parentId === parent._id)];
-        //   child.forEach((_child, j) => {
-        //     _child.srno = `${(i + 1)}.${(j + 1)}`
-        //     _child.detail = _child.productDetail?.productName
-        //     _child.qty = _child.qty * parent.qty
-        //     const subRows = []
-        //     const inventory = data.inventory?.filter((e) => e._id === _child._id);
-        //     inventory?.forEach((_inventory, l) => {
-        //       const transferFilter = transferAssets.filter(e => e.assetId === _inventory.inventory);
-        //       var isTransferAsset = false;
-        //       var transferData = {};
-        //       if (transferFilter.length) {
-        //         isTransferAsset = true
-        //         transferData = transferFilter[0]
-        //       }
-        //       const isPurchaseOrderAsset = purchaseOrderProduct.some(e => e._id === _inventory?.inventoryDetail?.purchaseOrder);
-        //       subRows.push({
-        //         ..._inventory,
-        //         srno: `${(i + 1)}.${(j + 1)}.${(l + 1)}`,
-        //         detail: _inventory?.inventoryDetail?.assetNumber,
-        //         type: "asset",
-        //         status: _inventory.inventoryDetail?.status,
-        //         manualStatus: _inventory.inventoryDetail?.manualStatus,
-        //         _id: _inventory.inventory,
-        //         isValid: _inventory.inventoryDetail?.manualStatus === INVENTORY_STATUS.reserved ? false : true,
-        //         isPurchaseOrderAsset: isPurchaseOrderAsset,
-        //         isTransferAsset: isTransferAsset,
-        //         transferData: transferData,
-        //         isSubleaseAsset: _inventory?.inventoryDetail?.subleaseAsset
-        //       })
-        //     })
-        //     _child.subRows = subRows;
-        //     _child.isValid = _child?.qty === subRows?.length ? true : false;
-        //     _child.isPurchaseOrder = purchaseOrderProduct?.some(e => e.productId === _child.materialId)
-        //     _child.isSublease = subleaseProduct?.some(e => e.materialId === _child.materialId)
-        //   })
-        //   if (child.filter(e => e.isValid === false).length > 0) {
-        //     parent.isValid = false
-        //   } else {
-        //     parent.isValid = true
-        //   }
-        //   parent.subRows = child;
-        // }
       });
 
       if (rows.filter(_rows => _rows.isValid === false).length > 0) {
@@ -346,7 +272,7 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
       }
 
       setRowsData(rows);
-      setSelectedProducts([])
+      setSelectedRecords([])
     }
     catch (error) {
       toastConfig.setToastConfig(error);
@@ -415,28 +341,15 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
       return <p>---</p>;
     }
     return <p>{row?.original?.assetAssignedQty} / {row?.original?.assetQty}</p>;
-    // if (row.original?.type === "product") {
-    //   if (row.subRows && row.subRows.filter(e => e.type === "asset")?.length > 0) {
-    //     return <p>{row.subRows.filter(e => e.type === "asset")?.length} / {row.original.qty}</p>
-    //   }
-    //   return <p>0 / {row.original.qty}</p>;
-    // }
-    // else if (row.original?.type === "package") {
-    //   const qty = row.original?.subRows?.reduce((sum, row) => row.qty + sum, 0);
-    //   const flatData = treeToFlatArray([{ ...row.original }], "subRows")
-    //   const getAssetsOnly = flatData.filter(f => f.type === "asset");
-    //   return <p>{getAssetsOnly.length} / {qty}</p>;
-    // }
-    // return "";
   }
 
   const handleAddSerializedAsset = (assets) => {
-    let data = [];
-    let flatArray = treeToFlatArray(selectedProducts, "subRows").filter(f => f.type === "product");
+    var data = [];
+    var flatArray = treeToFlatArray(selectedRecords, "subRows").filter(f => f.type === "product");
     flatArray = uniqBy(flatArray, '_id')
     flatArray?.forEach((e: any) => {
       if (e.type === "product") {
-        let qty = e.realAssetQty - e.assetAssignedQty;
+        let qty = e.realAssetQty - e.realAssetAssignedQty;
         while (qty) {
           const result = assets.filter(f => f.productId === e.materialId && !f.isCounted);
           if (result.length) {
@@ -451,13 +364,14 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
         }
       }
     })
+
     if (data.length) {
       setAdding(true)
       axiosInstance().post(`${rentalManagement.api}/${rentalManagementData._id}/inventory`, { "products": data })
         .then(({ data }) => {
           setAddSerializedAssetDialog({ open: false })
           fetchProductInventory()
-          setSelectedProducts([])
+          setSelectedRecords([])
           setAssetAssignedProduct([])
           setAdding(false)
           toastConfig.setToastConfig({
@@ -491,7 +405,7 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
   }
 
   useEffect(() => {
-    let flatArray = treeToFlatArray(selectedProducts, "subRows").filter(f => f.type === "product" && f.serializedProduct && f.realAssetQty > f.realAssetAssignedQty);
+    let flatArray = treeToFlatArray(selectedRecords, "subRows").filter(f => f.type === "product" && f.serializedProduct && f.realAssetQty > f.realAssetAssignedQty);
     flatArray = uniqBy(flatArray, '_id')
     const products = flatArray.map(m => { return { _id: m.materialId, unit: m.unit, assetsCount: m.realAssetQty - m.realAssetAssignedQty } })
     const uniqProduct = []
@@ -528,13 +442,13 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
         }
       }
     })
-    setAssetAssignedProduct(assetProduct)
-  }, [selectedProducts])
+    setAssetAssignedProduct([...assetProduct])
+  }, [selectedRecords])
 
   const disableAssignSerializedAssets = () => {
-    if (selectedProducts.length === 0)
+    if (selectedRecords.length === 0)
       return true;
-    const flatArray = treeToFlatArray(selectedProducts, "subRows").filter(f => f.type === "product" && f.realAssetQty > f.realAssetAssignedQty);
+    const flatArray = treeToFlatArray(selectedRecords, "subRows").filter(f => f.type === "product" && f.realAssetQty > f.realAssetAssignedQty);
     return flatArray.length === 0;
   }
 
@@ -603,9 +517,9 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
               >
                 {`Create ${routes.sublease.title}`}</MenuItem>}
             <MenuItem
-              disabled={(treeToFlatArray(selectedProducts, "subRows")?.filter(d => d.type === "asset" && d.status === INVENTORY_STATUS.reserved).length === 0)}
+              disabled={(treeToFlatArray(selectedRecords, "subRows")?.filter(d => d.type === "asset" && d.status === INVENTORY_STATUS.reserved).length === 0)}
               onClick={() => {
-                setDeleteData(treeToFlatArray(selectedProducts, "subRows")?.filter(d => d.type === "asset").map(d => d?.inventory))
+                setDeleteData(treeToFlatArray(selectedRecords, "subRows")?.filter(d => d.type === "asset").map(d => d?.inventory))
                 setShowConfirmBox(true)
                 closeActions()
               }}
@@ -641,7 +555,7 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
                 if (rowData.isSubleaseAsset) return "isSublease";
                 return "";
               }}
-              onSelect={setSelectedProducts}
+              onSelect={setSelectedRecords}
               childrenProperty="subRows"
               uniqueKey="_id"
               hideSelection={isOffline}
@@ -687,7 +601,7 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
         onClose={() => setOrderDialog(prevState => ({ ...prevState, open: false, type: "" }))}
         onSuccess={() => {
           setOrderDialog(({ open: false, products: [], type: "" }))
-          setSelectedProducts([])
+          setSelectedRecords([])
           fetchProductInventory()
           toastConfig.setToastConfig({
             open: true,
@@ -711,7 +625,7 @@ const SerializedAsset = ({ rentalManagementData, isTabletScreen, isSmallScreen, 
         onClose={() => setOrderDialog(prevState => ({ ...prevState, open: false, type: "" }))}
         onSuccess={() => {
           setOrderDialog(({ open: false, products: [], type: "" }))
-          setSelectedProducts([])
+          setSelectedRecords([])
           fetchProductInventory()
           toastConfig.setToastConfig({
             open: true,